+++ init.go
@@ -1,6 +1,7 @@
 package servermanager
 
 import (
+	"fmt"
 	"math/rand"
 	"os"
 	"os/signal"
@@ -44,6 +45,17 @@ func InitWithResolver(resolver *Resolver) error {
 			// ^C, handle it
 			if process.IsRunning() {
 				event := process.Event()
+
+				opts, err := store.LoadServerOptions()
+
+				if err == nil && opts.RestartEventOnServerManagerLaunch == 1 {
+					// save the event so it can be started again next time server-manager starts
+					err := store.UpsertLastRaceEvent(event)
+
+					if err != nil {
+						logrus.WithError(err).Error("Could not save last server event")
+					}
+				}
 
 				if event.IsChampionship() && !event.IsPractice() {
 					if err := championshipManager.StopActiveEvent(); err != nil {
@@ -102,6 +114,34 @@ func InitWithResolver(resolver *Resolver) error {
 			logrus.WithError(err).Error("Could not open search index")
 		}
 	}()
+
+	if opts.RestartEventOnServerManagerLaunch == 1 {
+		if lastEvent, err := store.LoadLastRaceEvent(); err == nil && lastEvent != nil {
+			var err error
+
+			switch a := lastEvent.(type) {
+			case *ActiveChampionship:
+				if !a.IsPractice() {
+					err = championshipManager.applyConfigAndStart(a)
+				} else {
+					err = raceManager.applyConfigAndStart(a)
+				}
+			case *ActiveRaceWeekend:
+				if !a.IsPractice() {
+					err = raceWeekendManager.applyConfigAndStart(a)
+				} else {
+					err = raceManager.applyConfigAndStart(a)
+				}
+			default:
+				fmt.Println("default")
+				err = raceManager.applyConfigAndStart(a)
+			}
+
+			if err != nil {
+				logrus.WithError(err).Errorf("Could not start last running event")
+			}
+		}
+	}
 
 	return nil
 }