+++ store.go
@@ -25,6 +25,9 @@ type Store interface {
 	// Live Timings
 	UpsertLiveTimingsData(*LiveTimingsPersistedData) error
 	LoadLiveTimingsData() (*LiveTimingsPersistedData, error)
+	UpsertLastRaceEvent(r RaceEvent) error
+	LoadLastRaceEvent() (RaceEvent, error)
+	ClearLastRaceEvent() error
 
 	UpsertLiveFrames([]string) error
 	ListPrevFrames() ([]string, error)